Clock: fix Tempo underflow bug

+++ software/src/HSClockManager.h
@@ -63,7 +63,7 @@ public:
         NR_OF_CLOCKS
     };
 
-    uint16_t tempo; // The set tempo, for display somewhere else
+    uint16_t tempo; // effective BPM, for display somewhere else
     uint16_t tempo_setting;
     uint32_t ticks_per_beat; // Based on the selected tempo in BPM
     bool running = 0; // Specifies whether the clock is running for interprocess communication
@@ -111,7 +111,7 @@ public:
      * This is approximate, because the arithmetical value is likely to be fractional, and we
      * need to live with a certain amount of imprecision here. So I'm not even rounding up.
      */
-    void SetTempoBPM(uint16_t bpm) {
+    void SetTempoBPM(int bpm) {
         bpm = constrain(bpm, CLOCK_TEMPO_MIN, CLOCK_TEMPO_MAX);
         ticks_per_beat = 1000000 / bpm;
         tempo_setting = tempo = bpm;
@@ -138,7 +138,7 @@ public:
     /* Gets the current tempo. This can be used between client processes, like two different
      * hemispheres.
      */
-    uint16_t GetTempo() {return tempo_setting;}
+    int GetTempo() {return tempo_setting;}
     float GetTempoFloat() {
       return 1000000.0f / ticks_per_beat;
     }
